+++ apollo-portal/src/test/java/com/ctrip/framework/apollo/portal/controller/ServerConfigControllerTest.java
@@ -18,14 +18,22 @@ package com.ctrip.framework.apollo.portal.controller;
 
 import com.ctrip.framework.apollo.portal.AbstractIntegrationTest;
 import com.ctrip.framework.apollo.portal.entity.po.ServerConfig;
+import com.ctrip.framework.apollo.portal.service.ServerConfigService;
 import org.junit.Assert;
 import org.junit.Test;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.jdbc.Sql;
 import org.springframework.web.client.HttpClientErrorException;
+
+import java.util.*;
+
 import static org.hamcrest.core.StringContains.containsString;
 import static org.junit.Assert.assertEquals;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.mockito.Mockito.*;
 
 /**
  * Created by kezhenxu at 2019/1/14 13:24.
@@ -34,7 +42,13 @@ import static org.hamcrest.MatcherAssert.assertThat;
  */
 @ActiveProfiles("skipAuthorization")
 public class ServerConfigControllerTest extends AbstractIntegrationTest {
+  @Mock
+  private ServerConfigService serverConfigService;
+  @InjectMocks
+  private ServerConfigController serverConfigController;
+
   @Test
+  @Sql(scripts = "/sql/cleanup.sql", executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD)
   public void shouldSuccessWhenParameterValid() {
     ServerConfig serverConfig = new ServerConfig();
     serverConfig.setKey("validKey");
@@ -75,5 +89,13 @@ public class ServerConfigControllerTest extends AbstractIntegrationTest {
           containsString("ServerConfig.Value cannot be blank")
       );
     }
+  }
+
+  @Test
+  public void testFindEmpty() {
+    when(serverConfigService.findAll()).thenReturn(new ArrayList<>());
+    List<ServerConfig> serverConfigList = serverConfigController.findAllServerConfig();
+    Assert.assertNotNull(serverConfigList);
+    Assert.assertEquals(0, serverConfigList.size());
   }
 }